Move code from zend_default_classes.c to zend_attributes.c
